+++ ledger/api/corporation/ledger.py
@@ -2,16 +2,15 @@ from typing import List
 
 from ninja import NinjaAPI
 
-from allianceauth.eveonline.models import EveCorporationInfo
-
 from ledger.api import schema
 from ledger.api.helpers import (
     get_corporation,
-    get_corps_members,
+    get_main_and_alts_all,
     get_main_and_alts_corporations,
 )
 from ledger.api.managers.ledger_manager import JournalProcess
 from ledger.hooks import get_extension_logger
+from ledger.models import CorporationAudit
 
 logger = get_extension_logger(__name__)
 
@@ -26,7 +25,7 @@ class LedgerApiEndpoints:
             tags=self.tags,
         )
         def get_corporation_ledger(request, corporation_id: int, year: int, month: int):
-            response, main = get_corporation(request, corporation_id)
+            response, _ = get_corporation(request, corporation_id)
 
             if not response:
                 return 403, "Permission Denied"
@@ -34,10 +33,11 @@ class LedgerApiEndpoints:
             if corporation_id == 0:
                 corporations = get_main_and_alts_corporations(request)
             else:
-                corporations = [main.corporation_id]
+                corporations = set()
+                corporations.add(corporation_id)
 
             # Create the Ledger
-            characters, chars_list = get_corps_members(corporations)
+            characters, chars_list = get_main_and_alts_all(corporations)
 
             ledger = JournalProcess(characters, year, month)
             output = ledger.corporation_ledger(chars_list)
@@ -51,7 +51,7 @@ class LedgerApiEndpoints:
         )
         def get_corporation_admin(request):
             if request.user.has_perm("ledger.admin_access"):
-                corporations = EveCorporationInfo.objects.all()
+                corporations = CorporationAudit.objects.visible_to(request.user)
             else:
                 return 403, "Permission Denied"
 
@@ -60,11 +60,12 @@ class LedgerApiEndpoints:
             for corporation in corporations:
                 # pylint: disable=broad-exception-caught
                 try:
-                    corporation_dict[corporation.corporation_id] = {
-                        "corporation_id": corporation.corporation_id,
-                        "corporation_name": corporation.corporation_name,
+                    corporation_dict[corporation.corporation.corporation_id] = {
+                        "corporation_id": corporation.corporation.corporation_id,
+                        "corporation_name": corporation.corporation.corporation_name,
                     }
-                except Exception:
+                except Exception as e:
+                    logger.debug(e)
                     continue
 
             output = []