+++ ledger/api/helpers.py
@@ -1,6 +1,6 @@
 from django.core.exceptions import ObjectDoesNotExist
 
-from allianceauth.eveonline.models import EveCharacter, EveCorporationInfo
+from allianceauth.eveonline.models import EveCharacter
 
 from ledger import app_settings, models
 from ledger.errors import LedgerImportError
@@ -57,61 +57,20 @@ def get_corp_models_and_string():
     return CorpMember
 
 
-def get_main_character(request, character_id):
-    """Get Main Character and check permissions"""
-    perms = True
-    if character_id == 0:
-        character_id = request.user.profile.main_character.character_id
-
-    main_char = EveCharacter.objects.select_related(
-        "character_ownership",
-        "character_ownership__user__profile",
-        "character_ownership__user__profile__main_character",
-    ).get(character_id=character_id)
-    # pylint: disable=broad-exception-caught
-    try:
-        main_char = main_char.character_ownership.user.profile.main_character
-    except Exception:
-        pass
-
-    # check access
-    visible = models.CharacterAudit.objects.visible_eve_characters(request.user)
-    if main_char not in visible:
-        account_chars = (
-            request.user.profile.main_character.character_ownership.user.character_ownerships.all()
-        )
-        account_char_ids = account_chars.values_list("character_id", flat=True)
-
-        logger.warning(
-            "%s Can See %s, requested %s",
-            request.user,
-            list(visible),
-            main_char.id,
-        )
-
-        if main_char.id not in account_char_ids:
-            perms = False
-
-    if not request.user.has_perm("ledger.basic_access"):
-        logger.warning(
-            "%s does not have Perm requested, Requested %s", request.user, main_char.id
-        )
-        perms = False
-
-    return perms, main_char
-
-
 def get_character(request, character_id):
     """Get Character and check permissions"""
     perms = True
     if character_id == 0:
         character_id = request.user.profile.main_character.character_id
 
-    main_char = EveCharacter.objects.select_related(
-        "character_ownership",
-        "character_ownership__user__profile",
-        "character_ownership__user__profile__main_character",
-    ).get(character_id=character_id)
+    try:
+        main_char = EveCharacter.objects.select_related(
+            "character_ownership",
+            "character_ownership__user__profile",
+            "character_ownership__user__profile__main_character",
+        ).get(character_id=character_id)
+    except ObjectDoesNotExist:
+        main_char = None
 
     # check access
     visible = models.CharacterAudit.objects.visible_eve_characters(request.user)
@@ -125,11 +84,15 @@ def get_corporation(request, corporation_id):
     perms = True
     if corporation_id == 0:
         corporation_id = request.user.profile.main_character.corporation_id
-
-    main_corp = EveCorporationInfo.objects.get(corporation_id=corporation_id)
+    try:
+        main_corp = models.CorporationAudit.objects.get(
+            corporation__corporation_id=corporation_id
+        )
+    except ObjectDoesNotExist:
+        main_corp = None
 
     # check access
-    visible = models.CorporationAudit.objects.visible_eve_corporations(request.user)
+    visible = models.CorporationAudit.objects.visible_to(request.user)
     if main_corp not in visible:
         perms = False
     return perms, main_corp
@@ -148,96 +111,64 @@ def get_alts_queryset(main_char):
         return EveCharacter.objects.filter(pk=main_char.pk)
 
 
-def get_corps_members(main_corp: EveCorporationInfo):
+def get_main_and_alts_all(main_corp: list):
     """Get all members for a corporation"""
     characters = {}
     chars_list = set()
-    corps_list = list(main_corp)
+    corps_list = main_corp
     corpmember = get_corp_models_and_string()
-    logger.debug("Main Corp: %s", main_corp)
-    corps = corpmember.objects.select_related("corpstats", "corpstats__corp").filter(
-        corpstats__corp__corporation_id__in=corps_list
+    corps = (
+        corpmember.objects.select_related("corpstats", "corpstats__corp")
+        .prefetch_related("corpstats")
+        .filter(corpstats__corp__corporation_id__in=corps_list)
+        .values_list("character_id", flat=True)
     )
-    for char in corps:
+    for character_id in corps:
         try:
-            char = EveCharacter.objects.get(character_id=char.character_id)
-        except EveCharacter.DoesNotExist:
-            pass
-        if char.character_id not in chars_list:
-            characters[char.character_id] = {"main": char, "alts": []}
-            chars_list.add(char.character_id)
-
-    return characters, chars_list
-
-
-def get_main_and_alts_all(corporations: list, corp_members=True):
-    """
-    Get all mains and their alts from given corporations
-
-    Args:
-    - corporations: corp `list`
-    - corp_members: add Corp Members
-
-    Returns - Dict(Queryset) & List
-    - `Dict`: Mains and Alts Queryset
-    - `List`: Character IDS
-    """
-    mains = {}
-    chars_list = set()
-    corpmember = get_corp_models_and_string()
-
-    linked_chars = EveCharacter.objects.filter(corporation_id__in=corporations)
-    linked_chars = linked_chars | EveCharacter.objects.filter(
-        character_ownership__user__profile__main_character__corporation_id__in=corporations
-    )
-
-    linked_chars = linked_chars.select_related(
-        "character_ownership", "character_ownership__user__profile__main_character"
-    ).prefetch_related("character_ownership__user__character_ownerships")
-    linked_chars = linked_chars.order_by("character_name")
-
-    for char in linked_chars:
-        try:
+            # Combine Alts with Main
+            char = (
+                EveCharacter.objects.select_related(
+                    "character_ownership",
+                    "character_ownership__user__profile__main_character",
+                )
+                .prefetch_related("character_ownership__user__character_ownerships")
+                .get(character_id=character_id)
+            )
             main = char.character_ownership.user.profile.main_character
             if main is not None:
                 linked_characters = (
-                    main.character_ownership.user.character_ownerships.all().exclude(
-                        character__character_id=main.character_id
-                    )
+                    main.character_ownership.user.character_ownerships.select_related(
+                        "character", "user"
+                    ).all()
                 )
-                if main.corporation_id in corporations:
+
+                linked_characters = linked_characters.exclude(
+                    character__character_id=main.character_id
+                )
+
+                if main.corporation_id in corps_list:
                     alts = [char.character for char in linked_characters]
-                    mains[main.character_id] = {
+                    characters[main.character_id] = {
                         "main": main,
                         "alts": alts,
                     }
+                    for alt in alts:
+                        chars_list.add(alt.character_id)
                 chars_list.add(main.character_id)
-                for alt in alts:
-                    chars_list.add(alt.character_id)
 
         except ObjectDoesNotExist:
-            pass
-
-    if corp_members:
-        corp = corpmember.objects.select_related("corpstats", "corpstats__corp").filter(
-            corpstats__corp__corporation_id__in=corporations
-        )
-
-        for char in corp:
-            try:
-                char = EveCharacter.objects.get(character_id=char.character_id)
-            except EveCharacter.DoesNotExist:
-                pass
-            if char.character_id not in chars_list:
-                mains[char.character_id] = {"main": char, "alts": []}
-                chars_list.add(char.character_id)
-    return mains, list(chars_list)
+            # Add Corp Members as main
+            characters[char.character_id] = {"main": char, "alts": []}
+            chars_list.add(char.character_id)
+    return characters, list(chars_list)
 
 
 def get_main_and_alts_corporations(request):
-    linked_characters = request.user.profile.main_character.character_ownership.user.character_ownerships.all().values_list(
-        "character_id", flat=True
-    )
+    linked_characters = request.user.profile.main_character.character_ownership.user.character_ownerships.select_related(
+        "character", "user"
+    ).all()
+
+    linked_characters = linked_characters.values_list("character_id", flat=True)
     chars = EveCharacter.objects.filter(id__in=linked_characters)
 
     corporations = set()