+++ ledger/api/managers/billboard_manager.py
@@ -1,10 +1,18 @@
 import math
+from collections import defaultdict
 from dataclasses import dataclass, field
-from datetime import datetime
 from decimal import Decimal
 
-from ledger import app_settings
-from ledger.api.managers.core_manager import LedgerData, LedgerDate, LedgerModels
+from django.db.models import Case, DecimalField, Q, Sum, When
+from django.db.models.functions import Coalesce, TruncDay, TruncMonth
+
+from ledger.api.helpers import convert_ess_payout
+from ledger.api.managers.core_manager import (
+    LedgerData,
+    LedgerDate,
+    LedgerFilter,
+    LedgerModels,
+)
 from ledger.hooks import get_extension_logger
 
 logger = get_extension_logger(__name__)
@@ -34,144 +42,143 @@ class BillboardLedger:
         }
         self.date_billboard = ["x"]
 
-    def aggregate_corp(self, journal, range_):
-        """Aggregate the journal entries for the corporation."""
-        total_bounty = 0
-        total_ess_payout = 0
-        for entry in journal:
-            if (self.date.month == 0 and entry["date"].month == range_) or (
-                self.date.month != 0 and entry["date"].day == range_
-            ):
+    def annotate_days(self):
+        if self.date.month == 0:
+            # Gruppieren nach Monat
+            annotations = {"period": TruncMonth("date")}
+        else:
+            # Gruppieren nach Tag
+            annotations = {"period": TruncDay("date")}
+
+        data_dict = defaultdict(lambda: defaultdict(int))
+        filters = LedgerFilter(self.chars)
+        donations_filter = filters.filter_donation & ~Q(first_party_id__in=self.chars)
+
+        if self.models.corp_journal:
+            corp_journal = (
+                self.models.corp_journal.annotate(**annotations)
+                .values("period")
+                .annotate(
+                    total_bounty=Sum(
+                        Case(
+                            When(filters.filter_bounty, then="amount"),
+                            default=0,
+                            output_field=DecimalField(),
+                        )
+                    ),
+                    total_ess=Sum(
+                        Case(
+                            When(filters.filter_ess, then="amount"),
+                            default=0,
+                            output_field=DecimalField(),
+                        )
+                    ),
+                )
+                .order_by("period")
+            )
+
+            for entry in corp_journal:
+                period = entry["period"].strftime("%Y-%m-%d")
                 if self.is_corp:
-                    if entry["ref_type"] == "bounty_prizes":
-                        total_bounty += entry["amount"]
-                if entry["ref_type"] == "ess_escrow_transfer":
-                    if self.is_corp:
-                        total_ess_payout += entry["amount"]
-                    else:
-                        total_ess_payout += (
-                            entry["amount"] / app_settings.LEDGER_CORP_TAX
-                        ) * (100 - app_settings.LEDGER_CORP_TAX)
-        return total_bounty, total_ess_payout
+                    data_dict[period]["total_bounty"] = entry["total_bounty"]
+                data_dict[period]["total_ess"] = entry["total_ess"]
 
-    def aggregate_char(self, journal, range_):
-        """Aggregate the journal entries for the character."""
-        total_bounty = 0
-        total_miscellaneous = 0
-        total_isk = 0
-        total_cost = 0
-        total_market_cost = 0
-        total_production_cost = 0
+                if not self.is_corp:
+                    # Convert the ESS Payout for Character
+                    data_dict[period]["total_ess"] = convert_ess_payout(
+                        data_dict[period]["total_ess"]
+                    )
 
-        for entry in journal:
-            if (self.date.month == 0 and entry["date"].month == range_) or (
-                self.date.month != 0 and entry["date"].day == range_
-            ):
-                # Bounty Filter
-                if entry["ref_type"] == "bounty_prizes":
-                    total_bounty += entry["amount"]
-                # Misc Filter
-                if (
-                    entry["ref_type"]
-                    in [
-                        "market_transaction",
-                        "contract_price_payment_corp",
-                        "contract_reward",
-                        "contract_price",
-                    ]
-                    and entry["amount"] > 0
-                ):
-                    total_miscellaneous += entry["amount"]
-                # Donation Filter
-                if (entry["ref_type"] == "player_donation") and entry["amount"] > 0:
-                    if entry["first_party_id"] not in self.chars:
-                        total_miscellaneous += entry["amount"]
-                # Total ISK
-                if entry["amount"] > 0:
-                    total_isk += entry["amount"]
-                else:
-                    total_cost += entry["amount"]
-                # Total Market
-                if entry["ref_type"] in [
-                    "market_escrow",
-                    "transaction_tax",
-                    "market_provider_tax",
-                    "brokers_fee",
-                ]:
-                    if entry["amount"] < 0:
-                        total_market_cost += entry["amount"]
-                # Production Cost
-                if entry["ref_type"] in ["industry_job_tax", "manufacturing"]:
-                    total_production_cost += entry["amount"]
-        return (
-            total_bounty,
-            total_miscellaneous,
-            total_isk,
-            total_cost,
-            total_market_cost,
-            total_production_cost,
-        )
+        if self.models.char_journal:
+            char_journal = (
+                self.models.char_journal.annotate(**annotations)
+                .values("period")
+                .annotate(
+                    total_bounty=Sum(
+                        Case(
+                            When(filters.filter_bounty, then="amount"),
+                            default=0,
+                            output_field=DecimalField(),
+                        )
+                    ),
+                    total_miscellaneous=Coalesce(
+                        Sum(
+                            "amount",
+                            filter=filters.filter_market
+                            | filters.filter_contract
+                            | donations_filter,
+                        ),
+                        0,
+                        output_field=DecimalField(),
+                    ),
+                    total_isk=Sum(
+                        Case(
+                            When(filters.filter_total, then="amount"),
+                            default=0,
+                            output_field=DecimalField(),
+                        )
+                    ),
+                    total_cost=Sum(
+                        Case(
+                            When(filters.filter_costs, then="amount"),
+                            default=0,
+                            output_field=DecimalField(),
+                        )
+                    ),
+                    total_market_cost=Sum(
+                        Case(
+                            When(filters.filter_market_cost, then="amount"),
+                            default=0,
+                            output_field=DecimalField(),
+                        )
+                    ),
+                    total_production_cost=Sum(
+                        Case(
+                            When(filters.filter_production, then="amount"),
+                            default=0,
+                            output_field=DecimalField(),
+                        )
+                    ),
+                )
+                .order_by("period")
+            )
 
-    def aggregate_mining(self, journal, range_):
-        """Aggregate the journal entries for the mining."""
-        total_mining = 0
-        for entry in journal:
-            if (self.date.month == 0 and entry["date"].month == range_) or (
-                self.date.month != 0 and entry["date"].day == range_
-            ):
-                total_mining += entry["total"] if entry["total"] else 0
-        return total_mining
+            for entry in char_journal:
+                period = entry["period"].strftime("%Y-%m-%d")
+                data_dict[period]["total_bounty"] = entry["total_bounty"]
+                data_dict[period]["total_miscellaneous"] = entry["total_miscellaneous"]
+                data_dict[period]["total_isk"] = entry["total_isk"]
+                data_dict[period]["total_cost"] = entry["total_cost"]
+                data_dict[period]["total_market_cost"] = entry["total_market_cost"]
+                data_dict[period]["total_production_cost"] = entry[
+                    "total_production_cost"
+                ]
 
-    # pylint: disable=too-many-branches
-    def process_day(
-        self, range_, corp_journal=None, char_journal=None, mining_journal=None
-    ):
-        """Process the day for the journal entries."""
-        date = datetime.now()
+        if self.models.mining_journal:
+            for entry in self.models.mining_journal.values("total", "date"):
+                period = entry["date"].strftime("%Y-%m-%d")
+                data_dict[period]["total_mining"] += entry["total"]
 
-        if self.date.monthly:
-            date = date.replace(day=1)
+        for period, data in data_dict.items():
+            # Main Data
+            self.data.total_bounty = data.get("total_bounty", 0)
+            self.data.total_ess_payout = data.get("total_ess", 0)
+            self.data.total_miscellaneous = data.get("total_miscellaneous", 0)
+            self.data.total_mining = data.get("total_mining", 0)
+            # Total
+            self.data.total_isk += data.get("total_isk", 0)
+            # Costs
+            self.data.total_cost += data.get("total_cost", 0)
+            self.data.total_market_cost += data.get("total_market_cost", 0)
+            self.data.total_production_cost += data.get("total_production_cost", 0)
 
-        if corp_journal:
-            total_bounty, total_ess = self.aggregate_corp(corp_journal, range_)
-            self.data.total_bounty = total_bounty
-            self.data.total_ess_payout = total_ess
-        if char_journal:
-            (
-                total_bounty,
-                total_miscellaneous,
-                total_isk,
-                total_cost,
-                total_market_cost,
-                total_production_cost,
-            ) = self.aggregate_char(char_journal, range_)
+            self.sum.sum_amount.append(int(self.data.total_bounty))
+            self.sum.sum_amount_ess.append(int(self.data.total_ess_payout))
+            if not self.is_corp:
+                self.sum.sum_amount_misc.append(int(self.data.total_miscellaneous))
+                self.sum.sum_amount_mining.append(int(self.data.total_mining))
 
-            self.data.total_bounty = total_bounty
-            self.data.total_miscellaneous = total_miscellaneous
-            # Wallet Charts
-            self.data.total_isk += total_isk
-            self.data.total_cost += total_cost
-            self.data.total_market_cost += total_market_cost
-            self.data.total_production_cost += total_production_cost
-
-        if mining_journal:
-            total_mining = self.aggregate_mining(mining_journal, range_)
-            self.data.total_mining = total_mining
-
-        # Add the totals to the respective lists
-        self.sum.sum_amount.append(int(self.data.total_bounty))
-        self.sum.sum_amount_ess.append(int(self.data.total_ess_payout))
-        if not self.is_corp:
-            self.sum.sum_amount_misc.append(int(self.data.total_miscellaneous))
-            self.sum.sum_amount_mining.append(int(self.data.total_mining))
-        # Add the date to the date list
-        self.date_billboard.append(
-            date.replace(
-                year=self.date.year, month=self.date.month, day=range_
-            ).strftime("%Y-%m-%d")
-            if not self.date.monthly
-            else date.replace(year=self.date.year, month=range_).strftime("%Y-%m")
-        )
+            self.date_billboard.append(period)
 
     def calculate_total_sum(self):
         self.sum.total_sum = sum(
@@ -282,36 +289,30 @@ class BillboardLedger:
     # Create the Billboard Char Ledger
     def billboard_char_ledger(self, chars: list):
         self.chars = chars
-        corp_journal_values = self.models.corp_journal.values(
-            "amount", "ref_type", "first_party_id", "second_party_id", "date"
-        )
-        char_journal_values = self.models.char_journal.values(
-            "amount", "ref_type", "first_party_id", "second_party_id", "date"
-        )
-        mining_journal_values = self.models.mining_journal.values("total", "date")
-        for day in self.date.day_checks:
-            self.process_day(
-                day, corp_journal_values, char_journal_values, mining_journal_values
-            )
 
+        # Greaters the Annotations
+        self.annotate_days()
+
+        # Calculate the Total Sums
         self.calculate_total_sum()
 
+        # Create Gauge Billboard
         rounded_percentages = self.calculate_percentages()
         self.generate_gauge_data(rounded_percentages)
 
+        # Create Ratting Bar Billboard
         self.generate_wallet_ratting_bar()
 
+        # Create Wallet Charts Billboard
         self.generate_wallet_charts_data()
 
         return self.billboard_dict
 
     # Create the Billboard Corp Ledger
-    def billboard_corp_ledger(self, corporation_dict, summary_dict_all):
-        corp_journal_values = self.models.corp_journal.values(
-            "amount", "ref_type", "first_party_id", "second_party_id", "date"
-        )
-        for day in self.date.day_checks:
-            self.process_day(day, corp_journal_values)
+    def billboard_corp_ledger(self, corporation_dict, summary_dict_all, chars_list):
+        self.chars = chars_list
+
+        self.annotate_days()
 
         self.calculate_total_sum()