+++ ledger/api/managers/core_manager.py
@@ -79,11 +79,13 @@ class LedgerFilterCore:
     filter_bounty: Q = field(init=False)
     filter_ess: Q = field(init=False)
     filter_mining: Q = field(init=False)
+    filter_total: Q = field(init=False)
 
     def __post_init__(self):
         self.filter_first_party = Q(first_party_id__in=self.char_id)
         self.filter_second_party = Q(second_party_id__in=self.char_id)
         self.filter_partys = self.filter_first_party | self.filter_second_party
+        self.filter_total = self.filter_partys & Q(amount__gt=0)
         self.filter_bounty = self.filter_second_party & Q(ref_type="bounty_prizes")
         self.filter_ess = self.filter_second_party & Q(ref_type="ess_escrow_transfer")
         self.filter_mining = (
@@ -111,6 +113,7 @@ class LedgerFilterCost(LedgerFilterCore):
         self.filter_production = self.filter_partys & Q(
             ref_type__in=["industry_job_tax", "manufacturing"]
         )
+        self.filter_costs = self.filter_partys & Q(amount__lt=0)
 
 
 @dataclass