+++ ledger/tests/test_api/test_corp_ledger_api.py
@@ -56,39 +56,27 @@ class ManageApiLedgerCorpEndpointsTest(TestCase):
         self.assertEqual(response.status_code, 200)
         self.assertEqual(response.json(), expected_data)
 
-    @patch("ledger.api.corporation.ledger.get_cache_stale")
-    @patch("ledger.api.corporation.ledger.IS_TESTING", False)
-    def test_get_corporation_ledger_api_cache_no_testing(self, mock_get_cache_stale):
+    def test_get_corporation_ledger_api_cache_no_testing(self):
         self.client.force_login(self.user)
         url = "/ledger/api/corporation/0/ledger/year/2024/month/3/"
-
-        mock_get_cache_stale.return_value = CorpmonthlyMarch
 
         response = self.client.get(url)
         expected_data = CorpmonthlyMarch
         self.assertEqual(response.status_code, 200)
         self.assertEqual(response.json(), expected_data)
 
-    @patch("ledger.api.corporation.ledger.get_cache_stale")
-    @patch("ledger.api.corporation.ledger.IS_TESTING", True)
-    def test_get_corporation_ledger_api_no_cache_testing(self, mock_get_cache_stale):
+    def test_get_corporation_ledger_api_no_cache_testing(self):
         self.client.force_login(self.user)
         url = "/ledger/api/corporation/0/ledger/year/2024/month/3/"
-
-        mock_get_cache_stale.return_value = False
 
         response = self.client.get(url)
         expected_data = CorpmonthlyMarch
         self.assertEqual(response.status_code, 200)
         self.assertEqual(response.json(), expected_data)
 
-    @patch("ledger.api.corporation.ledger.get_cache_stale")
-    @patch("ledger.api.corporation.ledger.IS_TESTING", False)
-    def test_get_corporation_ledger_api_no_cache_no_testing(self, mock_get_cache_stale):
+    def test_get_corporation_ledger_api_no_cache_no_testing(self):
         self.client.force_login(self.user)
         url = "/ledger/api/corporation/0/ledger/year/2024/month/3/"
-
-        mock_get_cache_stale.return_value = False
 
         response = self.client.get(url)
         expected_data = CorpmonthlyMarch
@@ -121,11 +109,9 @@ class ManageApiLedgerCorpEndpointsTest(TestCase):
 
         self.assertEqual(response.status_code, 403)
 
-    @patch("ledger.api.corporation.ledger.get_cache_stale")
-    def test_get_corporation_ledger_api_no_data(self, mock_get_cache_stale):
+    def test_get_corporation_ledger_api_no_data(self):
         self.client.force_login(self.user3)
         url = "/ledger/api/corporation/0/ledger/year/2024/month/3/"
-        mock_get_cache_stale.return_value = False
 
         response = self.client.get(url)