+++ ledger/tests/test_api/test_helpers.py
@@ -23,9 +23,9 @@ from app_utils.testing import (
 
 from ledger.api.helpers import (
     get_alts_queryset,
+    get_character,
     get_corp_models_and_string,
     get_main_and_alts_all,
-    get_main_character,
     get_models_and_string,
 )
 from ledger.errors import LedgerImportError
@@ -58,6 +58,7 @@ class TestApiHelpers(TestCase):
         cls.user4, _ = create_user_from_evecharacter(
             1004,
         )
+        cls.corp = EveCorporationInfo.objects.get(corporation_id=2001)
 
     def test_get_main_and_alts_all_no_corp(self):
         # given
@@ -66,7 +67,7 @@ class TestApiHelpers(TestCase):
         char = EveCharacter.objects.get(character_id=1001)
         self.user4.profile.main_character.delete()
         # when
-        data, _ = get_main_and_alts_all([2001], corp_members=False)
+        data, _ = get_main_and_alts_all([self.corp.corporation_id])
         # then
         self.assertEqual(data, {1001: {"main": char, "alts": []}})
 
@@ -77,7 +78,7 @@ class TestApiHelpers(TestCase):
         request.user = self.user
         corp_stats = CorpStats.objects.create(
             token=Token.objects.get(user=self.user),
-            corp=EveCorporationInfo.objects.get(corporation_id=2001),
+            corp=self.corp,
         )
         CorpMember.objects.create(
             character_id=1005, character_name="Gerthd", corpstats=corp_stats
@@ -87,7 +88,7 @@ class TestApiHelpers(TestCase):
             mains[char.character_id] = {"main": char, "alts": []}
         excepted_data = mains
         # when
-        data, _ = get_main_and_alts_all([2001], corp_members=True)
+        data, _ = get_main_and_alts_all([self.corp.corporation_id])
         # then
         self.assertEqual(data, excepted_data)
 
@@ -98,7 +99,7 @@ class TestApiHelpers(TestCase):
         request.user = self.user
         corp_stats = CorpStats.objects.create(
             token=Token.objects.get(user=self.user),
-            corp=EveCorporationInfo.objects.get(corporation_id=2001),
+            corp=self.corp,
         )
         CorpMember.objects.create(
             character_id=1005, character_name="Gerthd", corpstats=corp_stats
@@ -111,7 +112,7 @@ class TestApiHelpers(TestCase):
             mains[char.character_id] = {"main": char, "alts": []}
         excepted_data = mains
         # when
-        data, _ = get_main_and_alts_all([2001], corp_members=True)
+        data, _ = get_main_and_alts_all([self.corp.corporation_id])
         # then
         self.assertEqual(data, excepted_data)
 
@@ -136,7 +137,7 @@ class TestApiHelpers(TestCase):
 
         excepted_data = mains
         # when
-        data, _ = get_main_and_alts_all([2001], corp_members=True)
+        data, _ = get_main_and_alts_all([self.corp.corporation_id])
         # then
         self.assertEqual(data, excepted_data)
 
@@ -148,7 +149,7 @@ class TestApiHelpers(TestCase):
         add_character_to_user(
             self.user, EveCharacter.objects.get(character_id=1004), is_main=False
         )
-        data = get_main_character(request, 1004)
+        data = get_character(request, 1004)
         character = EveCharacter.objects.get(character_id=1001)
         # then
         self.assertEqual(data, (True, character))
@@ -158,7 +159,7 @@ class TestApiHelpers(TestCase):
         request = self.factory.get("/")
         request.user = self.user3
         # when
-        data = get_main_character(request, 1001)
+        data = get_character(request, 1001)
         character = EveCharacter.objects.get(character_id=1001)
         # then
         self.assertEqual(data, (False, character))
@@ -170,7 +171,7 @@ class TestApiHelpers(TestCase):
         mock_visible.return_value = []
         request.user = self.user
         # when
-        data = get_main_character(request, 1001)
+        data = get_character(request, 1001)
         character = EveCharacter.objects.get(character_id=1001)
         # then
         self.assertEqual(data, (True, character))