+++ ledger/tests/test_views/test_views.py
@@ -9,7 +9,8 @@ from app_utils.testdata_factories import UserMainFactory
 
 from ledger.models.general import General
 from ledger.views.character.character_ledger import character_ledger
-from ledger.views.pve import ledger_index, ratting_index
+from ledger.views.corporation.corporation_ledger import corporation_ledger
+from ledger.views.pve import ledger_index
 
 
 class TestViews(TestCase):
@@ -38,7 +39,9 @@ class TestViews(TestCase):
         self.assertEqual(response.status_code, HTTPStatus.OK)
 
     def test_corp_ledger_view(self):
-        request = self.factory.get(reverse("ledger:ledger_corp_index"))
+        request = self.factory.get(
+            reverse("ledger:corporation_ledger", kwargs={"corporation_pk": 0})
+        )
         request.user = self.user
-        response = ratting_index(request)
+        response = corporation_ledger(request, corporation_pk=0)
         self.assertEqual(response.status_code, HTTPStatus.OK)